requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
WTForms==3.2.1
alembic==1.16.4
email-validator==2.3.0
pytest==8.4.1
